+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+# 0.11.0
+- Methods which potentially mutate state are no longer const
+  - Base::writeCv
+  - FirmwareBase::eraseFirmware
+  - FirmwareBase::writeFirmware
+  - FirmwareBase::exitFirmware
+  - ZppBase::eraseZpp
+  - ZppBase::exitZpp
+
 # 0.10.0
 - ZPP-Valid-Query command is mandatory